schema tree CHANGE extend flags member in lys_node structures
Due to need of another flags, the flags member of lys_node had to be
extended from 8 to 14 bits. The structure itself shouldn't be
larger, since the 6 bits were taken from nacm member which now has
only 2 bits.

+++ src/parser.c
@@ -1324,11 +1324,11 @@ lyp_check_mandatory(struct lys_node *node)
 }
 
 int
-lyp_check_status(uint8_t flags1, struct lys_module *mod1, const char *name1,
-                 uint8_t flags2, struct lys_module *mod2, const char *name2,
+lyp_check_status(uint16_t flags1, struct lys_module *mod1, const char *name1,
+                 uint16_t flags2, struct lys_module *mod2, const char *name2,
                  const struct lys_node *node)
 {
-    uint8_t flg1, flg2;
+    uint16_t flg1, flg2;
 
     flg1 = (flags1 & LYS_STATUS_MASK) ? (flags1 & LYS_STATUS_MASK) : LYS_STATUS_CURR;
     flg2 = (flags2 & LYS_STATUS_MASK) ? (flags2 & LYS_STATUS_MASK) : LYS_STATUS_CURR;
